requirements/testing.txt
-r base.txt

bumpversion==0.5.3
flake8==3.5.0
freezegun==0.3.2
mock==1.3.0
pytest-cov==2.5.1
pycodestyle==2.4.0
